Changed file offsets to long ints, fixed signed/unsigned mistake

+++ qtv.h
@@ -4,7 +4,7 @@
 struct qtv_index
 {
 	int frame;
-	int offset;
+	long int offset;
 };
 
 struct qtv
